requirements.txt
flask
flask-socketio
eventlet
gunicorn
keras
python-dotenv
paho-mqtt
